Move all Grid logic into new component
