+++ TicTacToe.js
@@ -1,55 +1,23 @@
 import React, { useState } from 'react';
 import { Text, TouchableOpacity, View } from 'react-native';
 import { Container, Row, Col } from 'react-grid-system';
-import { times } from 'lodash'
+
+import Grid from './Grid'
 
 const TicTacToe = () => {
-  const gridHeight = 3
-  const gridWith = 3
-  const initialGrid = Array(gridHeight).fill().reduce((prev, _, rowIndex) => ({
-    ...prev,
-    [rowIndex + 1]: Array(gridWith).fill().reduce((prev, _, colIndex) => ({
-      ...prev,
-      [`col${colIndex+1}`]: null
-    }), {})
-  }), {})
-
-  const [columnState, setColumnState] = useState(initialGrid)
   const [isPlayerXTurn, setIsPlayerXTurn] = useState(true)
 
-  const cellStyle = {
-    backgroundColor: '#275DAD',
-    width: 100,
-    height: 100,
-    border: '1px solid black',
-    marginBottom: 16
-  }
-
-  const handleTurn = ({row, col }) => setColumnState(previousState => {
-    const state = {
-      ...previousState,
-      [row]: { ...previousState[row], [col]:  isPlayerXTurn ? 'X' : '0' }
-    }
-
+  const handleTurn = () => {
+    const previsPlayerXTurn = isPlayerXTurn
     setIsPlayerXTurn(!isPlayerXTurn)
 
-    return state
-  })
+    return previsPlayerXTurn ? 'X' : 'O'
+  }
 
   return (
     <View>
       <Container>
-        {times(gridHeight, (rowIndex) =>
-          <Row key={rowIndex}>
-            {times(gridWith, (colIndex) =>
-              <Col>
-                <TouchableOpacity style={cellStyle} onPress={() => handleTurn({ row: rowIndex + 1, col: `col${colIndex + 1}` })}>
-                  <Text>{columnState[rowIndex + 1][`col${colIndex + 1}`]}</Text>
-                </TouchableOpacity>
-              </Col>
-            )}
-          </Row>
-        )}
+        <Grid setCellValue={handleTurn} height={3} width={3} />
       </Container>
     </View>
   )
